requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic[email]==2.11.2
pydantic-settings==2.8.1
sqlalchemy==2.0.40
asyncpg==0.30.0
alembic==1.15.2
supabase==2.15.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1

# OCR
pytesseract==0.3.13
Pillow==11.1.0
pdf2image==1.17.0
numpy>=2.0
opencv-python-headless>=4.9.0
